Added spawn state and changed collision names
Now all entities will be on collision layer 2 "entities".
Initial state of state machine will be spawnstate and when it ends, it will assign collision layer to the entity, to allow detection even if they spawned within detection area.
Will need to update HitboxComponent to check for faction relationships for correct damage.